requirements.txt
manim
pdfminer.six
pypdf
pypdfium2
